requirements.txt
pillow
pyautogui
keyboard
opencv-python-headless
easyocr
pandas
numpy
colorama
requests
cryptography
pyinstaller